move statsd metrics stuff into it's own folder
