+++ cmd/statsd-logger/main.go
@@ -5,22 +5,22 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/catkins/statsd-logger"
+	"github.com/catkins/statsd-logger/metrics"
 )
 
 func main() {
 	shutdownChan := make(chan os.Signal)
 	signal.Notify(shutdownChan, syscall.SIGINT, syscall.SIGTERM, syscall.SIGKILL)
 
-	server, err := statsdLogger.New("0.0.0.0:8125")
+	metricsServer, err := metrics.NewServer(metrics.DefaultAddress)
 	if err != nil {
 		panic(err)
 	}
 
 	go func() {
-		server.Listen()
+		metricsServer.Listen()
 	}()
 
 	<-shutdownChan
-	server.Close()
+	metricsServer.Close()
 }
